udffsck: 32-bit size_t won't support devices > 4 GiB
Device size needs to be uint64_t because size_t on 32-bit systems can be
only 32 bits. As part of this fix try to switch to uniform terminology for
that value.

Among other things this fixes the ability to find AVDP descriptors (beyond
the one at LSN 256) on large filesystems.